CraftManager class, simplify inventory constructor
Craft is in special CraftManager class, not in World.
So does BlockManager, in which normal block duplication
bug is fixed.
Inventory now does not require shred to make things.
Add World::PutBlock, which just puts block, when
World::SetBlock puts block and activates it.
And corresponding Shred::PutBlock and Shred::SetBlock.